+++ jaguar/lib/src/http/params/import.dart
@@ -111,12 +111,3 @@ class QueryParams extends CastableStringMap {
 
   QueryParams.FromQueryParam(QueryParams param) : super(param);
 }
-
-/// Interface that must be implemented by all validatable objects
-abstract class Validatable {
-  /// Validates the object
-  ///
-  /// Shall throw an exception in case of validation failure.
-  /// The thrown exception must contain the reason for the failure
-  void validate();
-}