+++ jaguar/test/response/json/json.dart
@@ -47,7 +47,7 @@ void main() {
     test('ResponseProcessor.Podo', () async {
       await resty.get('http://localhost:10000/response_processor/podo').exact(
           statusCode: 200,
-          mimeType: MimeType.json,
+          mimeType: MimeTypes.json,
           body:
               r'{"name":"Jaguar","motto":["Speed","Simplicity","Extensiblity"]}');
     });
@@ -55,13 +55,13 @@ void main() {
     test('ResponseProcessor.List', () async {
       await resty
           .get('http://localhost:10000/response_processor/nums')
-          .exact(statusCode: 200, mimeType: MimeType.json, body: '[1,2,3]');
+          .exact(statusCode: 200, mimeType: MimeTypes.json, body: '[1,2,3]');
     });
 
     test('StrResponse.Podo', () async {
       await resty.get('http://localhost:10000/strresponse/podo').exact(
           statusCode: 200,
-          mimeType: MimeType.json,
+          mimeType: MimeTypes.json,
           body:
               r'{"name":"Jaguar","motto":["Speed","Simplicity","Extensiblity"]}');
     });
@@ -69,7 +69,7 @@ void main() {
     test('StrResponse.List', () async {
       await resty
           .get('http://localhost:10000/strresponse/nums')
-          .exact(statusCode: 200, mimeType: MimeType.json, body: '[1,2,3]');
+          .exact(statusCode: 200, mimeType: MimeTypes.json, body: '[1,2,3]');
     });
   });
 }